+++ app/javascript/helpers/functions.js
@@ -127,3 +127,10 @@ export function isCrawler() {
     ua.includes('ia_archiver')
   )
 }
+
+export function sanitizeMarkdown (desc) {
+  // open external and image links in new tab
+  desc = desc.replace(/<a(\s+)(href=['"]https?:\/\/|href=['"]\/image)/gi, '<a$1target="_blank" $2')
+  // desc = desc.replace(/\n/g, '<br>')
+  return desc
+}